requirements.txt
streamlit
pandas
numpy
plotly
scikit-learn
matplotlib
seaborn
tensorflow-cpu
